+++ connector/connect/server/src/main/scala/org/apache/spark/sql/connect/service/SparkConnectService.scala
@@ -19,30 +19,20 @@ package org.apache.spark.sql.connect.service
 
 import java.util.concurrent.TimeUnit
 
-import scala.annotation.tailrec
-import scala.collection.mutable.ArrayBuffer
-import scala.util.control.NonFatal
-
 import com.google.common.base.Ticker
 import com.google.common.cache.CacheBuilder
-import com.google.protobuf.{Any => ProtoAny}
-import com.google.rpc.{Code => RPCCode, ErrorInfo, Status => RPCStatus}
-import io.grpc.{Server, Status}
+import io.grpc.Server
 import io.grpc.netty.NettyServerBuilder
-import io.grpc.protobuf.StatusProto
 import io.grpc.protobuf.services.ProtoReflectionService
 import io.grpc.stub.StreamObserver
-import org.apache.commons.lang3.StringUtils
-import org.json4s.JsonDSL._
-import org.json4s.jackson.JsonMethods.{compact, render}
 
-import org.apache.spark.{SparkEnv, SparkException, SparkThrowable}
-import org.apache.spark.api.python.PythonException
+import org.apache.spark.SparkEnv
 import org.apache.spark.connect.proto
 import org.apache.spark.connect.proto.{AddArtifactsRequest, AddArtifactsResponse}
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.connect.config.Connect.{CONNECT_GRPC_BINDING_PORT, CONNECT_GRPC_MAX_INBOUND_MESSAGE_SIZE}
+
 
 /**
  * The SparkConnectService implementation.
@@ -54,80 +44,8 @@ import org.apache.spark.sql.connect.config.Connect.{CONNECT_GRPC_BINDING_PORT, C
  */
 class SparkConnectService(debug: Boolean)
     extends proto.SparkConnectServiceGrpc.SparkConnectServiceImplBase
-    with Logging {
-
-  private def allClasses(cl: Class[_]): Seq[Class[_]] = {
-    val classes = ArrayBuffer.empty[Class[_]]
-    if (cl != null && !cl.equals(classOf[java.lang.Object])) {
-      classes.append(cl) // Includes itself.
-    }
-
-    @tailrec
-    def appendSuperClasses(clazz: Class[_]): Unit = {
-      if (clazz == null || clazz.equals(classOf[java.lang.Object])) return
-      classes.append(clazz.getSuperclass)
-      appendSuperClasses(clazz.getSuperclass)
-    }
-
-    appendSuperClasses(cl)
-    classes.toSeq
-  }
-
-  private def buildStatusFromThrowable(st: Throwable): RPCStatus = {
-    val message = StringUtils.abbreviate(st.getMessage, 2048)
-    RPCStatus
-      .newBuilder()
-      .setCode(RPCCode.INTERNAL_VALUE)
-      .addDetails(
-        ProtoAny.pack(
-          ErrorInfo
-            .newBuilder()
-            .setReason(st.getClass.getName)
-            .setDomain("org.apache.spark")
-            .putMetadata("classes", compact(render(allClasses(st.getClass).map(_.getName))))
-            .build()))
-      .setMessage(if (message != null) message else "")
-      .build()
-  }
-
-  private def isPythonExecutionException(se: SparkException): Boolean = {
-    // See also pyspark.errors.exceptions.captured.convert_exception in PySpark.
-    se.getCause != null && se.getCause
-      .isInstanceOf[PythonException] && se.getCause.getStackTrace
-      .exists(_.toString.contains("org.apache.spark.sql.execution.python"))
-  }
-
-  /**
-   * Common exception handling function for the Analysis and Execution methods. Closes the stream
-   * after the error has been sent.
-   *
-   * @param opType
-   *   String value indicating the operation type (analysis, execution)
-   * @param observer
-   *   The GRPC response observer.
-   * @tparam V
-   * @return
-   */
-  private def handleError[V](
-      opType: String,
-      observer: StreamObserver[V]): PartialFunction[Throwable, Unit] = {
-    case se: SparkException if isPythonExecutionException(se) =>
-      logError(s"Error during: $opType", se)
-      observer.onError(
-        StatusProto.toStatusRuntimeException(buildStatusFromThrowable(se.getCause)))
-
-    case e: Throwable if e.isInstanceOf[SparkThrowable] || NonFatal.apply(e) =>
-      logError(s"Error during: $opType", e)
-      observer.onError(StatusProto.toStatusRuntimeException(buildStatusFromThrowable(e)))
-
-    case e: Throwable =>
-      logError(s"Error during: $opType", e)
-      observer.onError(
-        Status.UNKNOWN
-          .withCause(e)
-          .withDescription(StringUtils.abbreviate(e.getMessage, 2048))
-          .asRuntimeException())
-  }
+      with ServiceErrorHandling
+      with Logging {
 
   /**
    * This is the main entry method for Spark Connect and all calls to execute a plan.
